requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
joblib>=1.3.0
gunicorn>=21.0.0
